126
타이틀씬 글씨 임시 추가
monster ray 만드는중

+++ Engine_SOURCE/CameraScript.cpp
@@ -269,7 +269,7 @@ namespace dru
 				void* p = new int();
 				srand((int)p);
 				int randvalue = rand() % 5000000;
-				srand((int)p * CTimeMgr::DeltaTime() * randvalue);
+				srand((int)p * static_cast<UINT>(CTimeMgr::DeltaTime()) * randvalue);
 				float magnitude = mShakeParams.magnitude *
 					(1.f - mShakeTimer / mShakeParams.duration);
